WIP: pass fasta file input

+++ src/utils.rs
@@ -1,3 +1,3 @@
-pub fn percent(d: usize, n: usize) -> usize {
-    ((d as f64 / 100_f64) * n as f64).floor() as usize
+pub fn percent(d: f64, n: usize) -> usize {
+    ((d / 100_f64) * n as f64).floor() as usize
 }
